UnitTest: Twilio unit-test passes with local emulator.
